Geeklog2.2.2

+++ system/classes/Cache/FileSystem.php
@@ -193,7 +193,11 @@ class FileSystem implements CacheInterface
     {
         $fileName = $this->getFileName($key);
 
-        return @unlink($fileName);
+        if (is_readable($fileName)) {
+            return @unlink($fileName);
+        } else {
+            return true;
+        }
     }
 
     /**
@@ -208,5 +212,19 @@ class FileSystem implements CacheInterface
         clearstatcache();
 
         return @is_file($fileName) && @is_readable($fileName);
+    }
+
+    /**
+     * Return the timestamp of cached item
+     *
+     * @param  string $key
+     * @return int|false    the timestamp when the item exists, false otherwise
+     */
+    public function getAge($key)
+    {
+        $fileName = $this->getFileName($key);
+        clearstatcache();
+
+        return @is_file($fileName) && @is_readable($fileName) ? filemtime($fileName) : false;
     }
 }
